refactor: integrate into next-eval
