+++ apps/web/package.json
@@ -16,10 +16,7 @@
     "test:bench": "vitest bench"
   },
   "dependencies": {
-    "@next-eval/shared": "workspace:*",
-    "@next-eval/html-core": "workspace:*",
-    "@next-eval/evaluation-core": "workspace:*",
-    "@next-eval/llm-core": "workspace:*",
+    "@next-eval/next-eval": "workspace:*",
     "ai": "4.3.2",
     "react": "19.1.0",
     "react-dom": "19.1.0",
