+++ apps/web/scripts/update-mdr-ground-truth.ts
@@ -1,7 +1,7 @@
 import { readFile, readdir, writeFile } from "node:fs/promises";
 import { join } from "node:path";
 import { runMDRWithDetails } from "@/lib/utils/runMDR";
-import { slimHtml } from "@next-eval/html-core/utils/slimHtml";
+import { slimHtml } from "@next-eval/next-eval/html/utils/slimHtml";
 import { JSDOM } from "jsdom";
 
 /**
