--- apps/web/src/app/api/utils/constants.ts
@@ -1,1 +1,0 @@
-export const GEMINI_PRO_2_5_PREVIEW_03 = "gemini-2.5-pro-preview-03-25";
