--- packages/evaluation-core/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@next-eval/evaluation-core",
-  "version": "0.0.1",
-  "type": "module",
-  "exports": {
-    "./*": "./src/*.ts"
-  },
-  "scripts": {
-    "lint": "biome check --diagnostic-level=error",
-    "format": "biome format --fix",
-    "check-types": "tsc --noEmit"
-  },
-  "dependencies": {
-    "node-html-parser": "^7.0.1",
-    "munkres-js": "1.2.2",
-    "zod": "^3.22.4"
-  },
-  "devDependencies": {
-    "@next-eval/tsconfig": "workspace:*",
-    "@types/node": "22.14.0",
-    "typescript": "5.8.3",
-    "@types/mustache": "^4.2.6"
-  },
-  "packageManager": "bun@1.2.15"
-}
