--- packages/evaluation-core/src/types/munkres.d.ts
@@ -1,4 +1,0 @@
-declare module "munkres-js" {
-  function munkres(costMatrix: number[][]): [number, number][];
-  export = munkres;
-}
