--- packages/llm-core/package.json
@@ -1,26 +1,0 @@
-{
-  "name": "@next-eval/llm-core",
-  "version": "0.0.1",
-  "type": "module",
-  "exports": {
-    "./*": "./src/*.ts"
-  },
-  "scripts": {
-    "lint": "biome check --diagnostic-level=error",
-    "format": "biome format --fix",
-    "check-types": "tsc --noEmit"
-  },
-  "dependencies": {
-    "@next-eval/shared": "workspace:*",
-    "ai": "4.3.2",
-    "@ai-sdk/google": "1.2.8",
-    "@google/generative-ai": "^0.11.3",
-    "mustache": "4.2.0"
-  },
-  "devDependencies": {
-    "@next-eval/tsconfig": "workspace:*",
-    "@types/node": "22.14.0",
-    "typescript": "5.8.3"
-  },
-  "packageManager": "bun@1.2.15"
-}
