+++ packages/next-eval/src/evaluation/interfaces/EvaluationResult.ts
