+++ packages/next-eval/src/llm/prompts/system-llm-slim.md
