+++ packages/next-eval/src/shared/interfaces/HtmlResult.ts
