+++ packages/next-eval/src/shared/interfaces/LLMResponse.ts
