+++ packages/next-eval/tsconfig.json
@@ -1,0 +1,12 @@
+{
+  "extends": "@next-eval/tsconfig/library.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "baseUrl": ".",
+    "paths": {
+      "@next-eval/next-eval/*": ["./src/*"]
+    }
+  },
+  "include": ["src/**/*"],
+  "exclude": ["node_modules", "dist"]
+}
