--- packages/shared/tsconfig.json
@@ -1,12 +1,0 @@
-{
-  "extends": "@next-eval/tsconfig/library.json",
-  "compilerOptions": {
-    "outDir": "dist",
-    "baseUrl": ".",
-    "paths": {
-      "@next-eval/shared/*": ["./src/*"]
-    }
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
